Adjust download chunks
