+++ src/main.rs
@@ -6,59 +6,7 @@ use tokio::{
     io::{self, AsyncReadExt, AsyncWriteExt},
 };
 
-use futures::future::join_all;
-
 const TASK_NUM: u8 = 8;
-
-async fn download_chunks(downloader: Arc<Downloader>, video: Arc<Video>) {
-    let chunk_size = video.content_lenth / TASK_NUM as u64;
-    let mut range_list = vec![];
-    let mut start = 0;
-    let mut end = 0;
-
-    while end < video.content_lenth {
-        end += chunk_size;
-        if end > video.content_lenth {
-            end = video.content_lenth;
-        }
-        range_list.push((start, end));
-        start = end + 1;
-    }
-
-    let mut handler_list = vec![];
-    for (index, range) in range_list.into_iter().enumerate() {
-        println!("download chunk {} from {} to {}", index, range.0, range.1);
-        let downloader = downloader.clone();
-        let video = video.clone();
-        let handler =
-            tokio::spawn(async move { downloader.download_chunk(video, range, index as u8).await });
-        handler_list.push(handler);
-    }
-    join_all(handler_list).await;
-}
-
-async fn download_entry(bv: &str) -> Arc<Video> {
-    let downloader = Arc::new(Downloader::new().unwrap());
-
-    let video = downloader
-        .build_video(bv.to_string())
-        .await
-        .map_or_else(|e| panic!("{}", e), |v| Arc::new(v));
-
-    match fs::create_dir(bv).await {
-        Ok(_) => {
-            let clone = video.clone();
-            println!("download {} start, title: `{}`", clone.bv, clone.title);
-            if clone.content_lenth > 0 {
-                download_chunks(downloader, clone).await;
-            } else {
-                todo!()
-            }
-            return video;
-        }
-        Err(_) => panic!("create video dir failed"),
-    }
-}
 
 async fn create_file(filepath: String) -> io::Result<fs::File> {
     fs::OpenOptions::new()
@@ -107,8 +55,25 @@ async fn merge_chunk_files(video: Arc<Video>) -> () {
 #[tokio::main]
 async fn main() -> () {
     let bv = "BV1Q14y1L76r";
-    let video = download_entry(bv).await;
-    let _ = merge_chunk_files(video).await;
+    let downloader = Downloader::new().unwrap();
+    if let Ok(video) = downloader.build_video(bv.to_string()).await {
+        let video = Arc::new(video);
+        match fs::create_dir(bv).await {
+            Ok(_) => {
+                println!("download {} start, title: `{}`", video.bv, video.title);
+                let downloader = Arc::new(downloader);
+                if video.content_lenth > 0 {
+                    downloader.download_chunks(video.clone()).await.unwrap();
+                    let _ = merge_chunk_files(video.clone()).await;
+                } else {
+                    todo!()
+                }
+            }
+            Err(_) => panic!("create video dir failed"),
+        }
+    } else {
+        todo!()
+    }
 
     // let matches = clap::App::new("Bilibili Video Downloader")
     //     .version(clap::crate_version!())
@@ -143,14 +108,6 @@ mod tests {
 
     use super::*;
     use anyhow::Result;
-
-    #[tokio::test]
-    async fn test_download_video() -> Result<()> {
-        let bv = "bvid";
-        let video = download_entry(bv).await;
-        merge_chunk_files(video).await;
-        Ok(())
-    }
 
     #[tokio::test]
     async fn test_download_bullet() -> Result<()> {
